Implemented Pools Page,Vote Page,Vote Completed Page
Redesigned the app, redisigned the code for Pools Page so it renders the pools dinamically from the server. Did the same for the Vote Page after implementing it. Implemented the Vote Completed Page as well which dissapears after 1400 ms, so it's just a confirmation for the customer that his/her vote was successful then it redirects the customer to the Pools Page.